PR feedback: Extract `Tailwindcss::ProcessRunner` for process-running logic
