+++ lib/tailwindcss-rails.rb
@@ -5,3 +5,4 @@ require_relative "tailwindcss/version"
 require_relative "tailwindcss/engines"
 require_relative "tailwindcss/engine"
 require_relative "tailwindcss/commands"
+require_relative "tailwindcss/process_runner"
